requirements.txt
flask==2.0.1
werkzeug==2.0.2
python-dotenv==0.19.1
gunicorn==20.1.0
requests==2.26.0
watchdog==2.1.9
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
openai==1.3.0
python-dateutil==2.8.2
pytz==2023.3
sqlalchemy==1.4.46
pyodbc==4.0.39 